+++ app/wyzebridge/stream.py
@@ -1,3 +1,4 @@
+import json
 import time
 from subprocess import Popen, TimeoutExpired
 from threading import Thread
@@ -6,7 +7,7 @@ from typing import Any, Callable, Optional, Protocol
 from wyzebridge.config import MQTT_DISCOVERY, SNAPSHOT_INT, SNAPSHOT_TYPE
 from wyzebridge.ffmpeg import rtsp_snap_cmd
 from wyzebridge.logging import logger
-from wyzebridge.mqtt import update_preview
+from wyzebridge.mqtt import mqtt_cam_control, publish_message, update_preview
 from wyzebridge.rtsp_event import RtspEvent
 
 
@@ -43,10 +44,10 @@ class Stream(Protocol):
     def get_info(self, item: Optional[str] = None) -> dict:
         ...
 
-    def get_status(self) -> str:
+    def status(self) -> str:
         ...
 
-    def send_cmd(self, cmd: str) -> dict:
+    def send_cmd(self, cmd: str, value: str | dict = "") -> dict:
         ...
 
 
@@ -85,18 +86,6 @@ class StreamManager:
     def get_all_cam_info(self) -> dict:
         return {uri: s.get_info() for uri, s in self.streams.items()}
 
-    def start(self, uri: str) -> bool:
-        return stream.start() if (stream := self.get(uri)) else False
-
-    def stop(self, uri: str) -> bool:
-        return stream.stop() if (stream := self.get(uri)) else False
-
-    def enable(self, uri: str) -> bool:
-        return stream.enable() if (stream := self.get(uri)) else False
-
-    def disable(self, uri: str) -> bool:
-        return stream.disable() if (stream := self.get(uri)) else False
-
     def stop_all(self) -> None:
         logger.info(f"Stopping {self.total} stream{'s'[:self.total^1]}")
         self.stop_flag = True
@@ -107,14 +96,17 @@ class StreamManager:
         self.stop_flag = False
         if self.thread:
             self.thread.start()
+        mqtt = mqtt_cam_control(self.streams, self.send_cmd)
         logger.info(f"🎬 {self.total} stream{'s'[:self.total^1]} enabled")
-        event = RtspEvent(self)
+        event = RtspEvent(self.streams)
         while not self.stop_flag:
             mtx_health()
             event.read(timeout=1)
             cams = self.health_check_all()
             if cams and SNAPSHOT_TYPE == "rtsp":
                 self.snap_all(cams)
+        if mqtt:
+            mqtt.loop_stop()
         logger.info("Stream monitoring stopped")
 
     def monior_snapshots(self) -> None:
@@ -151,34 +143,37 @@ class StreamManager:
             stop_subprocess(self.rtsp_snapshots.get(cam))
             self.rtsp_snap_popen(cam, True)
 
-    def get_status(self, uri: str) -> str:
-        if self.stop_flag:
-            return "stopping"
-        return stream.get_status() if (stream := self.get(uri)) else "unavailable"
-
     def get_sse_status(self) -> dict:
-        return {uri: cam.get_status() for uri, cam in self.streams.items()}
+        return {uri: cam.status() for uri, cam in self.streams.items()}
 
-    def send_cmd(self, cam_name: str, cmd: str) -> dict:
+    def send_cmd(self, cam_name: str, cmd: str, payload: str | dict = "") -> dict:
         """
         Send a command directly to the camera and wait for a response.
 
         Parameters:
-        - cam_name (str): uri-friendly camera name to send command.
-        - cmd (str): The command to send. See wyzebridge.wyze_control.CAM_CMDS
-          for available commands.
+        - cam_name (str): uri-friendly name of the camera.
+        - cmd (str): The camera/tutk command to send.
+        - payload (str): value for the tutk command.
 
         Returns:
         - dictionary: Results that can be converted to JSON.
         """
-        resp = {"status": "error", "command": cmd}
+        resp = {"status": "error", "command": cmd, "payload": payload}
+
         if not (stream := self.get(cam_name)):
             return resp | {"response": "Camera not found"}
-        cam_resp = stream.send_cmd(cmd)
+
+        if cam_resp := stream.send_cmd(cmd, payload):
+            status = cam_resp.get("value") if cam_resp.get("status") == "success" else 0
+            if isinstance(status, dict):
+                status = json.dumps(status)
+            publish_message(f"{cam_name}/{cmd}", status)
         return cam_resp if "status" in cam_resp else resp | cam_resp
 
-    def rtsp_snap_popen(self, cam_name: str, interval: bool = False) -> Popen:
-        self.start(cam_name)
+    def rtsp_snap_popen(self, cam_name: str, interval: bool = False) -> Optional[Popen]:
+        if not (stream := self.get(cam_name)):
+            return
+        stream.start()
         ffmpeg = self.rtsp_snapshots.get(cam_name)
         if not ffmpeg or ffmpeg.poll() is not None:
             ffmpeg = Popen(rtsp_snap_cmd(cam_name, interval))
@@ -188,7 +183,8 @@ class StreamManager:
     def get_rtsp_snap(self, cam_name: str) -> bool:
         if not (stream := self.get(cam_name)) or stream.health_check() < 1:
             return False
-        ffmpeg = self.rtsp_snap_popen(cam_name)
+        if not (ffmpeg := self.rtsp_snap_popen(cam_name)):
+            return False
         try:
             if ffmpeg.wait(timeout=10) == 0:
                 return True